+++ CHANGELOG.md
@@ -29,6 +29,7 @@ BREAKING CHANGES
     to an infraction, slash them proportional to their stake at the time
   * Add REST endpoint to unrevoke a validator previously revoked for downtime
   * Add REST endpoint to retrieve liveness signing information for a validator
+* [types] renamed rational.Evaluate to rational.Round{Int64, Int}
 
 FEATURES
 * [gaiacli] You can now attach a simple text-only memo to any transaction, with the `--memo` flag